# docs/openapi/v1/api.yaml
openapi: 3.0.3
info:
  title: Shop API
  description: |
    REST API магазина бытовой техники.

    API управляет клиентами, товарами, поставщиками и изображениями товаров.
    Идентификаторы ресурсов имеют формат UUID, даты — ISO 8601. Ошибки
    возвращаются как JSON-объекты с текстом сообщения.
  version: 1.0.0
servers:
- url: http://localhost:{port}/api/v1
  description: Локальный сервер
  variables:
    port:
      default: '8080'
      description: Порт HTTP-сервера
...
paths:
...
  "/suppliers":
    post:
      tags:
      - Suppliers
      summary: Добавить поставщика
      operationId: createSupplier
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SupplierCreate"
            example:
              name: ООО БытТехника
              phone_number: "+7 495 123-45-67"
              address:
                country: Россия
                city: Москва
                street: Промышленная улица, 5
      responses:
        '201':
          description: Поставщик создан
          headers:
            Location:
              description: URL созданного поставщика
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Supplier"
        '400':
          "$ref": "#/components/responses/ValidationError"
        '409':
          "$ref": "#/components/responses/IdempotencyConflict"
        '500':
          "$ref": "#/components/responses/InternalError"
    get:
      tags:
      - Suppliers
      summary: Получить всех поставщиков
      description: Если поставщики отсутствуют, возвращается пустой массив.
      operationId: getSuppliers
      responses:
        '200':
          description: Список поставщиков
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Supplier"
              examples:
                result:
                  value:
                  - id: 7116b493-97ab-472f-8ae2-b34ed382619c
                    name: ООО БытТехника
                    phone_number: "+7 495 123-45-67"
                    address:
                      id: dcab51ee-44cf-4342-bcda-4b542ef3044b
                      country: Россия
                      city: Москва
                      street: Промышленная улица, 5
                empty:
                  value: []
        '500':
          "$ref": "#/components/responses/InternalError"
...
components:
  schemas:
    Address:
      type: object
      required:
      - id
      - country
      - city
      - street
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        country:
          type: string
          minLength: 1
          maxLength: 100
        city:
          type: string
          minLength: 1
          maxLength: 100
        street:
          type: string
          minLength: 1
          maxLength: 255
      example:
        id: 53c7c1b0-e2cd-4cc8-9689-d56145c54a9b
        country: Россия
        city: Москва
        street: Тверская улица, 10
    Error:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: Стабильный машиночитаемый код ошибки
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          description: Понятное человеку описание ошибки
    AddressCreate:
      type: object
      additionalProperties: false
      required:
      - country
      - city
      - street
      properties:
        country:
          type: string
          minLength: 1
          maxLength: 100
        city:
          type: string
          minLength: 1
          maxLength: 100
        street:
          type: string
          minLength: 1
          maxLength: 255
      example:
        country: Россия
        city: Москва
        street: Тверская улица, 10
...
    Supplier:
      type: object
      required:
      - id
      - name
      - phone_number
      - address
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          minLength: 1
          maxLength: 255
        phone_number:
          type: string
          minLength: 5
          maxLength: 30
        address:
          "$ref": "#/components/schemas/Address"
      example:
        id: 7116b493-97ab-472f-8ae2-b34ed382619c
        name: ООО БытТехника
        phone_number: "+7 495 123-45-67"
        address:
          id: dcab51ee-44cf-4342-bcda-4b542ef3044b
          country: Россия
          city: Москва
          street: Промышленная улица, 5
    SupplierCreate:
      type: object
      additionalProperties: false
      required:
      - name
      - phone_number
      - address
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        phone_number:
          type: string
          minLength: 5
          maxLength: 30
        address:
          "$ref": "#/components/schemas/AddressCreate"
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        Уникальный ключ операции. Повтор с тем же ключом и телом возвращает
        сохранённый результат; тот же ключ с другим телом приводит к `409`.
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 255
      example: 550e8400-e29b-41d4-a716-446655440000
...
  responses:
    IdempotencyConflict:
      description: Ключ идемпотентности уже использован для другого запроса
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
          example:
            code: IDEMPOTENCY_KEY_REUSED
            message: idempotency key is already used with a different request
    ValidationError:
      description: Тело запроса не прошло валидацию
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
          example:
            code: VALIDATION_ERROR
            message: validation failed
...
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
          example:
            code: INTERNAL_ERROR
            message: internal server error
...